float
changed values to float
relinked test button in hud test scene.

+++ GAM130Prototype/Assets/4 Play/UI/HudAssets/CompassScript.cs
@@ -6,14 +6,14 @@ public class CompassScript : MonoBehaviour
 {
 
     [Range (0,1200)]
-    public int lengthOfDay = 360;
+    public float lengthOfDay = 360;
     public float rotateAmount;
     public float interval;
     Vector3 dir;
     
     private void Update()
     {
-        interval = -(360 / lengthOfDay);  //rull rotation(360) / daylength = rotate interval
+        interval = -(360f / lengthOfDay);  //rull rotation(360) / daylength = rotate interval
         Debug.Log(interval);
         interval = interval * Time.deltaTime;
